chore: rename pkgs
