+++ pkg/proxyserver/controllers/manifests.go
