+++ pkg/proxyserver/operator/authentication/interface.go
