+++ pkg/proxyserver/operator/authentication/selfsigned/secret.go
