+++ pkg/proxyserver/operator/authentication/selfsigned/self_sign_certificate.go
